(MSmith) Tile System Setup- Small Demo Scene
I have setup a basic tile system to quickly paint levels, its kind of goofy but it works. Ive used the pirate assets ive found to create a small scene. (Demo Scene) Im havent set up all the available assets to be drag and drop yet, and im not sure what all of them are but i will work on that later!